finished tests for the wall and floor texture, will soon implement it into the bonus part

+++ bonus/src/raycasting/draw_walls.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   draw_walls.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: mbirou <mbirou@student.42.fr>              +#+  +:+       +#+        */
+/*   By: mbirou <manutea.birou@gmail.com>           +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/09/30 14:00:37 by mbirou            #+#    #+#             */
-/*   Updated: 2024/10/15 14:21:00 by mbirou           ###   ########.fr       */
+/*   Updated: 2024/10/16 09:30:50 by mbirou           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -67,7 +67,7 @@ void	cd_draw_c_f(t_game *game, int x)
 
 	y = -1;
 	mid = (game->graphic.height ) / 2.
-		+ (((float)game->graphic.height - 1.) / 2.) * game->player.pitch;
+		+ ((game->graphic.height - 1.) / 2.) * game->player.pitch;
 	while (++y < mid)
 		mlx_put_pixel(game->screen, x, y, game->graphic.c);
 	y --;
